Move functions to specific modules
Break things up a bit from the nene/core.py file and separate into more
focused modules.
